backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.3
python-dotenv>=1.0.1
pydantic-settings>=2.2.1